LayoutTests/accessibility/ios-simulator/spinbutton.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
<script>
var successfullyParsed = false;
if (window.testRunner)
   testRunner.dumpAsText();
</script>
</head>
<body>

<div id="fakeButton" tabindex=0 role="spinbutton" aria-label="spinner">button</div>
<input id="spinner" name="spinner" value="5" aria-valuenow="5" autocomplete="off" role="spinbutton">

<p id="description"></p>
<div id="console"></div>

<script>

    description("This test makes sure that spinbuttons appear.");

    if (window.accessibilityController) {

        var button = accessibilityController.accessibleElementById("fakeButton");
        shouldBe("button.description", "'AXLabel: spinner'");
        shouldBeFalse("button.isIgnored");
        
        // Test the aria-valuenow attribute on the spinbutton.
        var button2 = accessibilityController.accessibleElementById("spinner");
        shouldBe("button2.stringValue", "'AXValue: 5.00'");
    }

    successfullyParsed = true;
</script>

<script src="../../resources/js-test-post.js"></script>

</body>
</html>
